Refactor Collection component
