+++ src/pages/Collection.jsx
@@ -1,30 +1,17 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { Link, useLocation, useNavigate, useParams } from 'react-router-dom';
-import { useClient } from '../context/client-context';
-import { Typography, Grid, Button, Tabs, Tab } from '@mui/material';
-import PointCard from '../components/Points/PointCard';
-import ErrorNotifier from '../components/ToastNotifications/ErrorNotifier';
-import SimilarSerachfield from '../components/Points/SimilarSerachfield';
+import { Typography, Grid, Tabs, Tab } from '@mui/material';
 import { CenteredFrame } from '../components/Common/CenteredFrame';
 import Box from '@mui/material/Box';
 import { SnapshotsTab } from '../components/Snapshots/SnapshotsTab';
 import CollectionInfo from '../components/Collections/CollectionInfo';
-import { getErrorMessage } from '../lib/get-error-message';
+import PointsTabs from '../components/Points/PointsTabs';
 
 function Collection() {
-  const pageSize = 10;
-
   const { collectionName } = useParams();
-  const [points, setPoints] = useState(null);
-  const [offset, setOffset] = useState(null);
-  const [errorMessage, setErrorMessage] = useState(null);
-  const [conditions, setConditions] = useState([]);
-  const { client: qdrantClient } = useClient();
   const navigate = useNavigate();
   const location = useLocation();
   const [currentTab, setCurrentTab] = useState(location.hash.slice(1) || 'points');
-
-  const [nextPageOffset, setNextPageOffset] = useState(null);
 
   const handleTabChange = (event, newValue) => {
     if (typeof newValue !== 'string') {
@@ -33,101 +20,9 @@ function Collection() {
     setCurrentTab(newValue);
     navigate(`#${newValue}`);
   };
-
-  const onConditionChange = (conditions) => {
-    setOffset(null);
-    setConditions(conditions);
-    if (conditions.length === 0) {
-      setPoints({ points: [] });
-    }
-  };
-
-  const deletePoint = (collectionName, pointIds) => {
-    return qdrantClient.delete(collectionName, { points: pointIds }).catch((error) => {
-      console.log(error);
-      const message = getErrorMessage(error, { withApiKey: { apiKey: qdrantClient.getApiKey() } });
-      message && setErrorMessage(message);
-    });
-  };
-
-  useEffect(() => {
-    const getPoints = async () => {
-      if (conditions.length !== 0) {
-        const recommendationIds = [];
-        const filters = [];
-        let usingVector = null;
-        conditions.forEach((condition) => {
-          if (condition.type === 'id') {
-            recommendationIds.push(condition.value);
-            if (condition.using) {
-              usingVector = condition.using;
-            }
-          } else if (condition.type === 'payload') {
-            filters.push({ key: condition.key, match: { value: condition.value } });
-          }
-        });
-        try {
-          if (recommendationIds.length !== 0) {
-            const newPoints = await qdrantClient.recommend(collectionName, {
-              positive: recommendationIds,
-              limit: pageSize + (offset || 0),
-              with_payload: true,
-              with_vector: true,
-              using: usingVector,
-              filter: {
-                must: filters,
-              },
-            });
-            setNextPageOffset(newPoints.length);
-            setPoints({ points: newPoints });
-            setErrorMessage(null);
-          } else if (filters.length !== 0) {
-            const newPoints = await qdrantClient.scroll(collectionName, {
-              filter: {
-                must: filters,
-              },
-              limit: pageSize + (offset || 0),
-              with_payload: true,
-              with_vector: true,
-            });
-            setPoints({
-              points: [...(newPoints?.points || [])],
-            });
-            setNextPageOffset(newPoints?.next_page_offset);
-            setErrorMessage(null);
-          }
-        } catch (error) {
-          const message = getErrorMessage(error, { withApiKey: { apiKey: qdrantClient.getApiKey() } });
-          message && setErrorMessage(message);
-          setPoints({});
-        }
-      } else {
-        try {
-          const newPoints = await qdrantClient.scroll(collectionName, {
-            offset,
-            limit: pageSize,
-            with_vector: true,
-            with_payload: true,
-          });
-          setPoints({
-            points: [...(points?.points || []), ...(newPoints?.points || [])],
-          });
-          setNextPageOffset(newPoints?.next_page_offset);
-          setErrorMessage(null);
-        } catch (error) {
-          const message = getErrorMessage(error, { withApiKey: { apiKey: qdrantClient.getApiKey() } });
-          message && setErrorMessage(message);
-          setPoints({});
-        }
-      }
-    };
-    getPoints();
-  }, [collectionName, offset, conditions]);
-
   return (
     <>
       <CenteredFrame>
-        {errorMessage !== null && <ErrorNotifier {...{ message: errorMessage }} />}
         <Grid container maxWidth={'xl'} spacing={3}>
           <Grid xs={12} item>
             <Typography variant="h4">{collectionName}</Typography>
@@ -144,63 +39,11 @@ function Collection() {
             </Box>
           </Grid>
 
-          {currentTab === 'info' && (
-            <Grid xs={12} item>
-              <CollectionInfo collectionName={collectionName} />
-            </Grid>
-          )}
-          {currentTab === 'points' && (
-            <>
-              <Grid xs={12} item>
-                <SimilarSerachfield conditions={conditions} onConditionChange={onConditionChange} />
-              </Grid>
-
-              {errorMessage && (
-                <Grid xs={12} item textAlign={'center'}>
-                  <Typography>⚠ Error: {errorMessage}</Typography>
-                </Grid>
-              )}
-              {!points && !errorMessage && (
-                <Grid xs={12} item textAlign={'center'}>
-                  <Typography> 🔃 Loading...</Typography>
-                </Grid>
-              )}
-              {points && !errorMessage && points.points?.length === 0 && (
-                <Grid xs={12} item textAlign={'center'}>
-                  <Typography>📪 No Points are presents, {collectionName} is empty</Typography>
-                </Grid>
-              )}
-              {points &&
-                !errorMessage &&
-                points.points?.map((point) => (
-                  <Grid xs={12} item key={point.id}>
-                    <PointCard
-                      point={point}
-                      onConditionChange={onConditionChange}
-                      conditions={conditions}
-                      collectionName={collectionName}
-                      deletePoint={deletePoint}
-                    />
-                  </Grid>
-                ))}
-              <Grid xs={12} item textAlign={'center'}>
-                <Button
-                  variant="outlined"
-                  disabled={!points || !nextPageOffset}
-                  onClick={() => {
-                    setOffset(nextPageOffset);
-                  }}
-                >
-                  Load More
-                </Button>
-              </Grid>
-            </>
-          )}
-          {currentTab === 'snapshots' && (
-            <Grid xs={12} item>
-              <SnapshotsTab collectionName={collectionName} />
-            </Grid>
-          )}
+          <Grid xs={12} item>
+            {currentTab === 'info' && <CollectionInfo collectionName={collectionName} />}
+            {currentTab === 'points' && <PointsTabs collectionName={collectionName} />}
+            {currentTab === 'snapshots' && <SnapshotsTab collectionName={collectionName} />}
+          </Grid>
         </Grid>
       </CenteredFrame>
     </>
